Moved get_positive_integer function
